Speed up ability added cooldown works

+++ Assets/prefabs/Zombie/AIController.cs
@@ -36,7 +36,7 @@ public class AIController : MonoBehaviour
         }
     }
 
-    private void HealthChanged(int newValue, int oldValue, int maxValue, GameObject Causer)
+    private void HealthChanged(float newValue, float oldValue, float maxValue, GameObject Causer)
     {
         PerceptionUpdated(true, Causer);
         StartForgetingHurt(Causer);
